Renaming everything to PSR-7 session
