+++ docs/limitations.md
@@ -1,10 +1,10 @@
-## StorageLessSession Limitations
+## PSR7Session Limitations
 
-StorageLessSession has a few limitations derived from its design.
+PSR7Session has a few limitations derived from its design.
 
 #### Cannot store private information in the session
 
-StorageLessSession stores session data in cookies in an unencrypted JWT token.
+PSR7Session stores session data in cookies in an unencrypted JWT token.
 The fact that the token is unencrypted means that all the information in the
 session is also available in read-only mode to the user agent.
 
@@ -18,9 +18,9 @@ sessions may be read by various processes.
 #### Sessions cannot be invalidated
 
 There is no way to (securely) manually invalidate a session just via
-StorageLessSession.
+PSR7Session.
 
-By default, StorageLessSession does not assign any identifier to sessions,
+By default, PSR7Session does not assign any identifier to sessions,
 nor identifies sessions at all: it just verifies the session signature to
 validate the author of its contents.
 
@@ -28,10 +28,10 @@ If you want to manually lock out a particular user agent that logged in with
 a certain session cookie, then you will have to design a mechanism for that.
 You will need to identify (and attach identifiers to) session cookies, and
 then manually block clients with those identifiers in the session cookie.
-Note that this approach also defeats the benefits of StorageLessSession,
+Note that this approach also defeats the benefits of PSR7Session,
 therefore you may want to just use traditional sessions.
 
-This limitation is also why StorageLessSession should only be used with secure
+This limitation is also why PSR7Session should only be used with secure
 (TLS) HTTPS connections: if any session is spoofed, there is no way to lock
 out an attacker.
 
@@ -43,21 +43,21 @@ from the user agent in every HTTP request.
 
 #### Race conditions with highly concurrent HTTP requests writing to session
 
-Since StorageLessSession uses the [`SetCookie`](https://en.wikipedia.org/wiki/HTTP_cookie#Setting_a_cookie)
+Since PSR7Session uses the [`SetCookie`](https://en.wikipedia.org/wiki/HTTP_cookie#Setting_a_cookie)
 header to write data to the user-agent, exploiting the user-agent as a storage,
 it is not safe to use it for highly concurrent write operations.
 
-The idea around StorageLessSession is that a session is not supposed to be
+The idea around PSR7Session is that a session is not supposed to be
 an actual storage for transient client information, but rather be used for
 the concerns of authentication, authorization and eventually and for validation
 concerns such as CSRF-token validation.
 
 If you want to store frequently-updated or concurrently-updated information
-inside a session, then StorageLessSession is likely not fitting your use-case.
+inside a session, then PSR7Session is likely not fitting your use-case.
 
 #### Limit on the amount of data stored in a session
 
-StorageLessSession session size limitations are directly related with the
+PSR7Session session size limitations are directly related with the
 [size of the cookies](http://www.ietf.org/rfc/rfc2965.txt) allowed on the
 user-agent.
 While RFC2965 encourages allowing arbitrary-length cookies, this sadly
